Added the base code for isolating the table from the video
Started adding the code which will allow us to remove all the extra scenery from the video, leaving just the table. Didn't quite work however as it was cutting out bits which it shouldnt have. Same issue when I drew in the bounds. Looks like its an issue with the points we get from our click.