Refactored WorldState
Changed all the methods from individual types to world objects. Added a lot more to ObjectPosition and refactored to remove Goal and Pitch so that they are now just ObjectPosisions. The moveable object interface is now gone, and replaced with a MovingObject class which instantiates a Vector2 and an ObjectPosition.